requirements.txt
attrs
click
numpy
tqdm
